requirements.txt
amqp==5.0.7
asgiref==3.4.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.1
cffi==1.15.0
channels==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Django==3.2.10
django-countries==7.2.1
django-environ==0.8.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.2
phonenumbers==8.12.39
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==21.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
